Added support for multiple tokens.
If we want to use the /dooropen command from multiple slack teams, we have to support multiple access tokens.
config.txt can now store a list of hashed access tokens separated by a "," the /dooropen endpoint will check the whole list of access tokens and allow the request, if one of them matches.